+++ pkg/ingester/ingester_v2_test.go
@@ -27,6 +27,7 @@ import (
 	"github.com/prometheus/prometheus/pkg/labels"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/tsdb"
+	"github.com/prometheus/prometheus/tsdb/chunkenc"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
 	"github.com/stretchr/testify/require"
@@ -37,6 +38,8 @@ import (
 	"github.com/weaveworks/common/user"
 	"google.golang.org/grpc"
 
+	"github.com/cortexproject/cortex/pkg/chunk/encoding"
+	"github.com/cortexproject/cortex/pkg/cortexpb"
 	"github.com/cortexproject/cortex/pkg/ingester/client"
 	"github.com/cortexproject/cortex/pkg/ring"
 	cortex_tsdb "github.com/cortexproject/cortex/pkg/storage/tsdb"
@@ -697,7 +700,7 @@ func Test_Ingester_v2LabelNames(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), "test")
 
 	for _, series := range series {
-		req, _, _ := mockWriteRequest(series.lbls, series.value, series.timestamp)
+		req, _, _, _ := mockWriteRequest(t, series.lbls, series.value, series.timestamp)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -741,7 +744,7 @@ func Test_Ingester_v2LabelValues(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), "test")
 
 	for _, series := range series {
-		req, _, _ := mockWriteRequest(series.lbls, series.value, series.timestamp)
+		req, _, _, _ := mockWriteRequest(t, series.lbls, series.value, series.timestamp)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -860,7 +863,7 @@ func Test_Ingester_v2Query(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), "test")
 
 	for _, series := range series {
-		req, _, _ := mockWriteRequest(series.lbls, series.value, series.timestamp)
+		req, _, _, _ := mockWriteRequest(t, series.lbls, series.value, series.timestamp)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -1154,7 +1157,7 @@ func Test_Ingester_v2MetricsForLabelMatchers(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), "test")
 
 	for _, series := range fixtures {
-		req, _, _ := mockWriteRequest(series.lbls, series.value, series.timestamp)
+		req, _, _, _ := mockWriteRequest(t, series.lbls, series.value, series.timestamp)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -1283,7 +1286,15 @@ func createIngesterWithSeries(t testing.TB, userID string, numSeries int, timest
 
 func TestIngester_v2QueryStream(t *testing.T) {
 	// Create ingester.
-	i, err := prepareIngesterWithBlocksStorage(t, defaultIngesterTestConfig(), nil)
+	cfg := defaultIngesterTestConfig()
+
+	// change stream type in runtime.
+	var streamType QueryStreamType
+	cfg.StreamTypeFn = func() QueryStreamType {
+		return streamType
+	}
+
+	i, err := prepareIngesterWithBlocksStorage(t, cfg, nil)
 	require.NoError(t, err)
 	require.NoError(t, services.StartAndAwaitRunning(context.Background(), i))
 	defer services.StopAndAwaitTerminated(context.Background(), i) //nolint:errcheck
@@ -1296,7 +1307,7 @@ func TestIngester_v2QueryStream(t *testing.T) {
 	// Push series.
 	ctx := user.InjectOrgID(context.Background(), userID)
 	lbls := labels.Labels{{Name: labels.MetricName, Value: "foo"}}
-	req, _, expectedResponse := mockWriteRequest(lbls, 123000, 456)
+	req, _, expectedResponseSamples, expectedResponseChunks := mockWriteRequest(t, lbls, 123000, 456)
 	_, err = i.v2Push(ctx, req)
 	require.NoError(t, err)
 
@@ -1317,7 +1328,7 @@ func TestIngester_v2QueryStream(t *testing.T) {
 	require.NoError(t, err)
 	defer c.Close()
 
-	s, err := c.QueryStream(ctx, &client.QueryRequest{
+	queryRequest := &client.QueryRequest{
 		StartTimestampMs: 0,
 		EndTimestampMs:   200000,
 		Matchers: []*client.LabelMatcher{{
@@ -1325,22 +1336,56 @@ func TestIngester_v2QueryStream(t *testing.T) {
 			Name:  model.MetricNameLabel,
 			Value: "foo",
 		}},
-	})
-	require.NoError(t, err)
-
-	count := 0
-	var lastResp *client.QueryStreamResponse
-	for {
-		resp, err := s.Recv()
-		if err == io.EOF {
-			break
-		}
-		require.NoError(t, err)
-		count += len(resp.Timeseries)
-		lastResp = resp
 	}
-	require.Equal(t, 1, count)
-	require.Equal(t, expectedResponse, lastResp)
+
+	samplesTest := func(t *testing.T) {
+		s, err := c.QueryStream(ctx, queryRequest)
+		require.NoError(t, err)
+
+		count := 0
+		var lastResp *client.QueryStreamResponse
+		for {
+			resp, err := s.Recv()
+			if err == io.EOF {
+				break
+			}
+			require.NoError(t, err)
+			require.Zero(t, len(resp.Chunkseries)) // No chunks expected
+			count += len(resp.Timeseries)
+			lastResp = resp
+		}
+		require.Equal(t, 1, count)
+		require.Equal(t, expectedResponseSamples, lastResp)
+	}
+
+	chunksTest := func(t *testing.T) {
+		s, err := c.QueryStream(ctx, queryRequest)
+		require.NoError(t, err)
+
+		count := 0
+		var lastResp *client.QueryStreamResponse
+		for {
+			resp, err := s.Recv()
+			if err == io.EOF {
+				break
+			}
+			require.NoError(t, err)
+			require.Zero(t, len(resp.Timeseries)) // No samples expected
+			count += len(resp.Chunkseries)
+			lastResp = resp
+		}
+		require.Equal(t, 1, count)
+		require.Equal(t, expectedResponseChunks, lastResp)
+	}
+
+	streamType = QueryStreamDefault
+	t.Run("default", samplesTest)
+
+	streamType = QueryStreamSamples
+	t.Run("samples", samplesTest)
+
+	streamType = QueryStreamChunks
+	t.Run("chunks", chunksTest)
 }
 
 func TestIngester_v2QueryStreamManySamples(t *testing.T) {
@@ -1359,7 +1404,7 @@ func TestIngester_v2QueryStreamManySamples(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), userID)
 
 	const samplesCount = 100000
-	samples := make([]client.Sample, 0, samplesCount)
+	samples := make([]cortexpb.Sample, 0, samplesCount)
 
 	for i := 0; i < samplesCount; i++ {
 		samples = append(samples, client.Sample{
@@ -1437,15 +1482,118 @@ func TestIngester_v2QueryStreamManySamples(t *testing.T) {
 	require.Equal(t, 10000+50000+samplesCount, totalSamples)
 }
 
-func writeRequestSingleSeries(lbls labels.Labels, samples []client.Sample) *client.WriteRequest {
-	req := &client.WriteRequest{
-		Source: client.API,
+func TestIngester_v2QueryStreamManySamplesChunks(t *testing.T) {
+	// Create ingester.
+	cfg := defaultIngesterTestConfig()
+	cfg.StreamChunksWhenUsingBlocks = true
+
+	i, err := prepareIngesterWithBlocksStorage(t, cfg, nil)
+	require.NoError(t, err)
+	require.NoError(t, services.StartAndAwaitRunning(context.Background(), i))
+	defer services.StopAndAwaitTerminated(context.Background(), i) //nolint:errcheck
+
+	// Wait until it's ACTIVE.
+	test.Poll(t, 1*time.Second, ring.ACTIVE, func() interface{} {
+		return i.lifecycler.GetState()
+	})
+
+	// Push series.
+	ctx := user.InjectOrgID(context.Background(), userID)
+
+	const samplesCount = 1000000
+	samples := make([]cortexpb.Sample, 0, samplesCount)
+
+	for i := 0; i < samplesCount; i++ {
+		samples = append(samples, cortexpb.Sample{
+			Value:       float64(i),
+			TimestampMs: int64(i),
+		})
 	}
 
-	ts := client.TimeSeries{}
-	ts.Labels = client.FromLabelsToLabelAdapters(lbls)
+	// 100k samples in chunks use about 154 KiB,
+	_, err = i.v2Push(ctx, writeRequestSingleSeries(labels.Labels{{Name: labels.MetricName, Value: "foo"}, {Name: "l", Value: "1"}}, samples[0:100000]))
+	require.NoError(t, err)
+
+	// 1M samples in chunks use about 1.51 MiB,
+	_, err = i.v2Push(ctx, writeRequestSingleSeries(labels.Labels{{Name: labels.MetricName, Value: "foo"}, {Name: "l", Value: "2"}}, samples))
+	require.NoError(t, err)
+
+	// 500k samples in chunks need 775 KiB,
+	_, err = i.v2Push(ctx, writeRequestSingleSeries(labels.Labels{{Name: labels.MetricName, Value: "foo"}, {Name: "l", Value: "3"}}, samples[0:500000]))
+	require.NoError(t, err)
+
+	// Create a GRPC server used to query back the data.
+	serv := grpc.NewServer(grpc.StreamInterceptor(middleware.StreamServerUserHeaderInterceptor))
+	defer serv.GracefulStop()
+	client.RegisterIngesterServer(serv, i)
+
+	listener, err := net.Listen("tcp", "localhost:0")
+	require.NoError(t, err)
+
+	go func() {
+		require.NoError(t, serv.Serve(listener))
+	}()
+
+	// Query back the series using GRPC streaming.
+	c, err := client.MakeIngesterClient(listener.Addr().String(), defaultClientTestConfig())
+	require.NoError(t, err)
+	defer c.Close()
+
+	s, err := c.QueryStream(ctx, &client.QueryRequest{
+		StartTimestampMs: 0,
+		EndTimestampMs:   samplesCount + 1,
+
+		Matchers: []*client.LabelMatcher{{
+			Type:  client.EQUAL,
+			Name:  model.MetricNameLabel,
+			Value: "foo",
+		}},
+	})
+	require.NoError(t, err)
+
+	recvMsgs := 0
+	series := 0
+	totalSamples := 0
+
+	for {
+		resp, err := s.Recv()
+		if err == io.EOF {
+			break
+		}
+		require.NoError(t, err)
+		require.True(t, len(resp.Chunkseries) > 0) // No empty messages.
+
+		recvMsgs++
+		series += len(resp.Chunkseries)
+
+		for _, ts := range resp.Chunkseries {
+			for _, c := range ts.Chunks {
+				ch, err := encoding.NewForEncoding(encoding.Encoding(c.Encoding))
+				require.NoError(t, err)
+				require.NoError(t, ch.UnmarshalFromBuf(c.Data))
+
+				totalSamples += ch.Len()
+			}
+		}
+	}
+
+	// As ingester doesn't guarantee sorting of series, we can get 2 (100k + 500k in first, 1M in second)
+	// or 3 messages (100k or 500k first, 1M second, and 500k or 100k last).
+
+	require.True(t, 2 <= recvMsgs && recvMsgs <= 3)
+	require.Equal(t, 3, series)
+	require.Equal(t, 100000+500000+samplesCount, totalSamples)
+}
+
+func writeRequestSingleSeries(lbls labels.Labels, samples []cortexpb.Sample) *cortexpb.WriteRequest {
+	req := &cortexpb.WriteRequest{
+		Source: cortexpb.API,
+	}
+
+	ts := cortexpb.TimeSeries{}
+	ts.Labels = cortexpb.FromLabelsToLabelAdapters(lbls)
 	ts.Samples = samples
-	req.Timeseries = append(req.Timeseries, client.PreallocTimeseries{TimeSeries: &ts})
+	req.Timeseries = append(req.Timeseries, cortexpb.PreallocTimeseries{TimeSeries: &ts})
 
 	return req
 }
@@ -1463,9 +1611,20 @@ func (m *mockQueryStreamServer) Context() context.Context {
 	return m.ctx
 }
 
-func BenchmarkIngester_v2QueryStream(b *testing.B) {
+func BenchmarkIngester_v2QueryStream_Samples(b *testing.B) {
+	benchmarkV2QueryStream(b, false)
+}
+
+func BenchmarkIngester_v2QueryStream_Chunks(b *testing.B) {
+	benchmarkV2QueryStream(b, true)
+}
+
+func benchmarkV2QueryStream(b *testing.B, streamChunks bool) {
+	cfg := defaultIngesterTestConfig()
+	cfg.StreamChunksWhenUsingBlocks = streamChunks
+
 	// Create ingester.
-	i, err := prepareIngesterWithBlocksStorage(b, defaultIngesterTestConfig(), nil)
+	i, err := prepareIngesterWithBlocksStorage(b, cfg, nil)
 	require.NoError(b, err)
 	require.NoError(b, services.StartAndAwaitRunning(context.Background(), i))
 	defer services.StopAndAwaitTerminated(context.Background(), i) //nolint:errcheck
@@ -1515,36 +1674,58 @@ func BenchmarkIngester_v2QueryStream(b *testing.B) {
 	}
 }
 
-func mockWriteRequest(lbls labels.Labels, value float64, timestampMs int64) (*client.WriteRequest, *client.QueryResponse, *client.QueryStreamResponse) {
-	samples := []client.Sample{
+func mockWriteRequest(t *testing.T, lbls labels.Labels, value float64, timestampMs int64) (*client.WriteRequest, *client.QueryResponse, *client.QueryStreamResponse, *client.QueryStreamResponse) {
+	samples := []cortexpb.Sample{
 		{
 			TimestampMs: timestampMs,
 			Value:       value,
 		},
 	}
 
-	req := client.ToWriteRequest([]labels.Labels{lbls}, samples, nil, client.API)
+	req := cortexpb.ToWriteRequest([]labels.Labels{lbls}, samples, nil, client.API)
 
 	// Generate the expected response
 	expectedQueryRes := &client.QueryResponse{
-		Timeseries: []client.TimeSeries{
+		Timeseries: []cortexpb.TimeSeries{
 			{
-				Labels:  client.FromLabelsToLabelAdapters(lbls),
+				Labels:  cortexpb.FromLabelsToLabelAdapters(lbls),
 				Samples: samples,
 			},
 		},
 	}
 
-	expectedQueryStreamRes := &client.QueryStreamResponse{
-		Timeseries: []client.TimeSeries{
+	expectedQueryStreamResSamples := &client.QueryStreamResponse{
+		Timeseries: []cortexpb.TimeSeries{
 			{
-				Labels:  client.FromLabelsToLabelAdapters(lbls),
+				Labels:  cortexpb.FromLabelsToLabelAdapters(lbls),
 				Samples: samples,
 			},
 		},
 	}
 
-	return req, expectedQueryRes, expectedQueryStreamRes
+	chunk := chunkenc.NewXORChunk()
+	app, err := chunk.Appender()
+	require.NoError(t, err)
+	app.Append(timestampMs, value)
+	chunk.Compact()
+
+	expectedQueryStreamResChunks := &client.QueryStreamResponse{
+		Chunkseries: []client.TimeSeriesChunk{
+			{
+				Labels: cortexpb.FromLabelsToLabelAdapters(lbls),
+				Chunks: []client.Chunk{
+					{
+						StartTimestampMs: timestampMs,
+						EndTimestampMs:   timestampMs,
+						Encoding:         int32(encoding.PrometheusXorChunk),
+						Data:             chunk.Bytes(),
+					},
+				},
+			},
+		},
+	}
+
+	return req, expectedQueryRes, expectedQueryStreamResSamples, expectedQueryStreamResChunks
 }
 
 func prepareIngesterWithBlocksStorage(t testing.TB, ingesterCfg Config, registerer prometheus.Registerer) (*Ingester, error) {
@@ -1940,7 +2121,7 @@ func TestIngester_invalidSamplesDontChangeLastUpdateTime(t *testing.T) {
 	sampleTimestamp := int64(model.Now())
 
 	{
-		req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, sampleTimestamp)
+		req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, sampleTimestamp)
 		_, err = i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -1955,7 +2136,7 @@ func TestIngester_invalidSamplesDontChangeLastUpdateTime(t *testing.T) {
 
 	// Push another sample to the same metric and timestamp, with different value. We expect to get error.
 	{
-		req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 1, sampleTimestamp)
+		req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 1, sampleTimestamp)
 		_, err = i.v2Push(ctx, req)
 		require.Error(t, err)
 	}
@@ -2249,7 +2430,7 @@ func Test_Ingester_v2UserStats(t *testing.T) {
 	ctx := user.InjectOrgID(context.Background(), "test")
 
 	for _, series := range series {
-		req, _, _ := mockWriteRequest(series.lbls, series.value, series.timestamp)
+		req, _, _, _ := mockWriteRequest(t, series.lbls, series.value, series.timestamp)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -2294,7 +2475,7 @@ func Test_Ingester_v2AllUserStats(t *testing.T) {
 	})
 	for _, series := range series {
 		ctx := user.InjectOrgID(context.Background(), series.user)
-		req, _, _ := mockWriteRequest(series.lbls, series.value, series.timestamp)
+		req, _, _, _ := mockWriteRequest(t, series.lbls, series.value, series.timestamp)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -2541,7 +2722,7 @@ func verifyCompactedHead(t *testing.T, i *Ingester, expected bool) {
 
 func pushSingleSampleWithMetadata(t *testing.T, i *Ingester) {
 	ctx := user.InjectOrgID(context.Background(), userID)
-	req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, util.TimeToMillis(time.Now()))
+	req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, util.TimeToMillis(time.Now()))
 	req.Metadata = append(req.Metadata, &client.MetricMetadata{MetricFamilyName: "test", Help: "a help for metric", Unit: "", Type: client.COUNTER})
 	_, err := i.v2Push(ctx, req)
 	require.NoError(t, err)
@@ -2549,7 +2730,7 @@ func pushSingleSampleWithMetadata(t *testing.T, i *Ingester) {
 
 func pushSingleSampleAtTime(t *testing.T, i *Ingester, ts int64) {
 	ctx := user.InjectOrgID(context.Background(), userID)
-	req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, ts)
+	req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, ts)
 	_, err := i.v2Push(ctx, req)
 	require.NoError(t, err)
 }
@@ -2692,7 +2873,7 @@ func TestIngesterNotDeleteUnshippedBlocks(t *testing.T) {
 	// Push some data to create 3 blocks.
 	ctx := user.InjectOrgID(context.Background(), userID)
 	for j := int64(0); j < 5; j++ {
-		req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, j*chunkRangeMilliSec)
+		req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, j*chunkRangeMilliSec)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -2719,7 +2900,7 @@ func TestIngesterNotDeleteUnshippedBlocks(t *testing.T) {
 
 	// Add more samples that could trigger another compaction and hence reload of blocks.
 	for j := int64(5); j < 6; j++ {
-		req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, j*chunkRangeMilliSec)
+		req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, j*chunkRangeMilliSec)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -2747,7 +2928,7 @@ func TestIngesterNotDeleteUnshippedBlocks(t *testing.T) {
 
 	// Add more samples that could trigger another compaction and hence reload of blocks.
 	for j := int64(6); j < 7; j++ {
-		req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, j*chunkRangeMilliSec)
+		req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, j*chunkRangeMilliSec)
 		_, err := i.v2Push(ctx, req)
 		require.NoError(t, err)
 	}
@@ -2793,7 +2974,7 @@ func TestIngesterPushErrorDuringForcedCompaction(t *testing.T) {
 	require.True(t, db.casState(active, forceCompacting))
 
 	// Ingestion should fail with a 503.
-	req, _, _ := mockWriteRequest(labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, util.TimeToMillis(time.Now()))
+	req, _, _, _ := mockWriteRequest(t, labels.Labels{{Name: labels.MetricName, Value: "test"}}, 0, util.TimeToMillis(time.Now()))
 	ctx := user.InjectOrgID(context.Background(), userID)
 	_, err = i.v2Push(ctx, req)
 	require.Equal(t, httpgrpc.Errorf(http.StatusServiceUnavailable, wrapWithUser(errors.New("forced compaction in progress"), userID).Error()), err)